mynewsproject/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
coverage==7.3.2
Django==4.2.6
django-debug-toolbar==4.2.0
django-extensions==3.2.3
Faker==12.0.1
h11==0.14.0
idna==3.4
mixer==7.2.2
outcome==1.3.0.post0
Pillow==10.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
wsproto==1.2.0
